move compiler detection tests
